requirements.txt
pandas>=2.2
numpy>=1.26
scikit-learn>=1.5
matplotlib>=3.9
seaborn>=0.13
umap-learn>=0.5
networkx>=3.3
jupyterlab>=4.2
ipywidgets>=8.1
streamlit
plotly
textblob
